Split pattern intefaces file
