+++ src/core/interfaces/pattern/index.ts
@@ -1,0 +1,4 @@
+// src/core/interfaces/pattern/index.ts
+export * from "./events";
+export * from "./tracks";
+export * from "./core";
